chore: move tls config from config.rs to tls/server_config.rs
